refactor: update import paths to use aliasing for consistency across components
