+++ apps/client/src/components/layouts/content-layout/footer.tsx
@@ -4,7 +4,7 @@ import TwitterLogo from "@/assets/icon-twitter.svg?react";
 import Logo from "@/assets/logo.svg?react";
 import { Container } from "@/components/ui/container";
 import { Link } from "react-router-dom";
-import { NavLinks } from "../nav-links";
+import { NavLinks } from "@/components/layouts/nav-links";
 import { paths } from "@/config/paths";
 
 export const Footer = () => {
