+++ apps/client/src/components/layouts/content-layout/index.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from "react-router-dom";
-import { Footer } from "./footer";
-import { Navbar } from "./nav-bar";
+import { Footer } from "@/components/layouts/content-layout/footer";
+import { Navbar } from "@/components/layouts/content-layout/nav-bar";
 
 export const ContentLayout = () => {
   return (
