+++ apps/client/src/features/categories/components/category-nav-dropdown/index.tsx
@@ -2,7 +2,7 @@ import Icon from "@/assets/icon-arrow-right.svg?react";
 import { getCategoriesQueryOptions } from "@/features/categories/api/get-categories";
 import { useQueryClient, useSuspenseQuery } from "@tanstack/react-query";
 import { Link } from "react-router-dom";
-import { getProductsByCategoryQueryOptions } from "../../../products/api/get-products";
+import { getProductsByCategoryQueryOptions } from "@/features/products/api/get-products";
 
 const CategoryNavDropdown = ({
   clickHandler,
