+++ apps/client/src/features/products/components/product-card/product-new-indicator.tsx
@@ -1,4 +1,4 @@
-import { UseProductCardContext } from "./index";
+import { UseProductCardContext } from "@/features/products/components/product-card/index";
 import { cn } from "@/lib/cn";
 
 export default function ProductNewIndicator(props: { classes?: string }) {
