+++ apps/client/src/main.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import { App } from "./app";
+import { App } from "@/app";
 import "./index.css";
 
 createRoot(document.getElementById("root")!).render(
